# packages/api-client/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aperture API",
        "version": "0.1.0",
        "description": "Aperture's HTTP API. This schema is the single source of truth for the frontend's types: it generates packages/api-client, which is never hand-edited."
    },
    "paths": {
...
        "/api/users/{username}": {
            "get": {
                "operationId": "users_profile",
                "description": "One account's public profile, with its counts.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/users/{username}/block": {
            "post": {
                "operationId": "users_block",
                "description": "Block an account, severing follows both ways.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "403": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            },
            "delete": {
                "operationId": "users_unblock",
                "description": "Unblock. Does not restore the follows that blocking severed.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/users/{username}/follow": {
            "post": {
                "operationId": "users_follow",
                "description": "Follow, or request to follow a private account.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FollowState"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            },
            "delete": {
                "operationId": "users_unfollow",
                "description": "Unfollow, or withdraw a pending request.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FollowState"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/users/{username}/respond": {
            "post": {
                "operationId": "users_respond_to_request",
                "description": "Approve or decline a pending follow request.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RespondToRequestRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RespondToRequestRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RespondToRequestRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "403": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/users/me": {
            "get": {
                "operationId": "users_me",
                "description": "The signed-in user.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "description": "No response body"
                    }
                }
            },
            "patch": {
                "operationId": "users_update_me",
                "description": "Update your own profile.",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUpdateProfileRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUpdateProfileRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUpdateProfileRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/users/requests": {
            "get": {
                "operationId": "users_follow_requests",
                "description": "Pending follow requests for your account.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FollowRequest"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/search": {
            "get": {
                "operationId": "users_search",
                "description": "Find accounts by username or display name.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Query",
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserList"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/session": {
            "post": {
                "operationId": "session_create",
                "description": "Sign in with email and password.",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "No response body"
                    }
                }
            },
            "delete": {
                "operationId": "session_destroy",
                "description": "Sign out. Idempotent.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CurrentUser": {
                "type": "object",
                "description": "The signed-in user seeing their own account. Adds the private bits.",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "username": {
                        "type": "string",
                        "readOnly": true,
                        "description": "Case-insensitively unique. Shown in metadata position."
                    },
                    "display_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "avatar_media_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "readOnly": true
                    },
                    "bio": {
                        "type": "string",
                        "readOnly": true
                    },
                    "is_private": {
                        "type": "boolean",
                        "readOnly": true,
                        "description": "Private accounts turn follows into requests pending approval."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "readOnly": true
                    }
                },
                "required": [
                    "avatar_media_id",
                    "bio",
                    "created_at",
                    "display_name",
                    "email",
                    "id",
                    "is_private",
                    "username"
                ]
            },
            "FollowRequest": {
                "type": "object",
                "description": "A pending request, from the perspective of the account being asked.",
                "properties": {
                    "follower": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "created_at",
                    "follower"
                ]
            },
            "FollowState": {
                "type": "object",
                "description": "What the follow button should read after acting.",
                "properties": {
                    "follow_state": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FollowStateEnum"
                            }
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "follow_state"
                ]
            },
            "FollowStateEnum": {
                "enum": [
                    "none",
                    "pending",
                    "accepted"
                ],
                "type": "string",
                "description": "* `none` - none\n* `pending` - pending\n* `accepted` - accepted"
            },
...
            "LoginRequest": {
                "type": "object",
                "description": "Credentials. Email is the login identifier here, not username.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "writeOnly": true,
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
            "PatchedUpdateProfileRequest": {
                "type": "object",
                "properties": {
                    "display_name": {
                        "type": "string",
                        "maxLength": 60
                    },
                    "bio": {
                        "type": "string",
                        "maxLength": 300
                    },
                    "is_private": {
                        "type": "boolean"
                    }
                }
            },
...
            "Profile": {
                "type": "object",
                "description": "A profile header.\n\nCounts come from the `counters` app, never from a `COUNT(*)` — a follower\ncount on a popular account is a sequential scan over millions of rows, and\nit renders on every profile view.\n\n`follow_state` is what the button reads: absent, `pending` while a private\naccount considers the request, or `accepted`.",
                "properties": {
                    "user": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "readOnly": true
                    },
                    "post_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "follower_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "following_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "follow_state": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FollowStateEnum"
                            }
                        ],
                        "readOnly": true
                    },
                    "is_self": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "can_view_posts": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "can_view_posts",
                    "follow_state",
                    "follower_count",
                    "following_count",
                    "is_self",
                    "post_count",
                    "user"
                ]
            },
            "RespondToRequestRequest": {
                "type": "object",
                "properties": {
                    "accept": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "accept"
                ]
            },
...
            "UserList": {
                "type": "object",
                "properties": {
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "users"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "aperture_session"
            }
        }
    }
}